+++ YX/Models/BaseMotorModel.cs
@@ -31,20 +31,20 @@ namespace YX.Models
     /// <summary>
     /// 额定电压（V）
     /// </summary>
-    [Range(0.1, double.MaxValue, ErrorMessage = "电压必须大于0")]
-    public double Voltage { get; set; }
+    [Range(0.1, (double)decimal.MaxValue, ErrorMessage = "电压必须大于0")]
+    public decimal Voltage { get; set; }
 
     /// <summary>
     /// 电机效率（%）
     /// </summary>
     [Range(0, 100, ErrorMessage = "效率需在0-100之间")]
-    public double MotorEfficiency { get; set; }
+    public decimal MotorEfficiency { get; set; }
 
     /// <summary>
     /// 最大效率点位置（%额定负载）
     /// </summary>
     [Range(0, 100, ErrorMessage = "最大效率点需在0-100之间")]
-    public double MaxEfficiencyLoadRatio { get; set; }
+    public decimal MaxEfficiencyLoadRatio { get; set; }
 
 #region 特性参数（使用 MotorDataPoint 封装）
 /// <summary>
